Refactor data generation script to enhance functionality and structure
- Expanded data generation to include items for sales and returns.
- Implemented realistic customer registration dates based on first purchase.
- Improved logging for better traceability during data generation.
- Updated functions to generate sales and returns with detailed itemization.
- Added status updates for sales based on return processing.
- Organized output CSV files into a dedicated 'data/' directory.
- Ensured data consistency and integrity across generated datasets.
